requirements.txt
django-ckeditor==5.4.0
django-taggit==0.22.2
pillow==4.0.0